Refactor ( architecture): Move bot-related code to a dedicated bot module, reorganize utility functions, and implement a listener structure for event handling. Remove obsolete files and ensure user context management with decorators. Update notification service and background tasks for improved functionality and error handling.
