+++ src/app/core/bot/bot_utils.py
